1.3.2-beta

+++ RaspberrySharp/IO/GeneralPurpose/ProcessorPinExtensionMethods.cs
@@ -20,7 +20,7 @@ namespace RaspberrySharp.IO.GeneralPurpose
         {
             return ((Enum.GetValues(typeof(ProcessorPin)) as ProcessorPin[]) ?? new ProcessorPin[0])
                 .Distinct()
-                .Where(p => (pins & (ProcessorPins)((uint)1 << (int)p)) != ProcessorPins.None)
+                .Where(p => (pins & (ProcessorPins)((ulong)1 << (int)p)) != ProcessorPins.None)
                 .ToArray();
         }
